DEV: Extract image orientation spec helpers
