+++ spec/lib/upload_creator_spec.rb
@@ -1,58 +1,10 @@
 # frozen_string_literal: true
 
-require "chunky_png"
 require "file_store/s3_store"
 
 RSpec.describe UploadCreator do
   fab!(:user)
   fab!(:admin)
-
-  def with_jpeg_orientation(source_path:, orientation:)
-    exif =
-      "Exif\0\0".b + "II*\0".b + [8].pack("V") + [1].pack("v") + [0x0112].pack("v") +
-        [3].pack("v") + [1].pack("V") + [orientation].pack("v") + "\0\0".b + [0].pack("V")
-    marker = "\xFF\xE1".b + [exif.bytesize + 2].pack("n") + exif
-    source = File.binread(source_path)
-    oriented_file = Tempfile.new(["oriented-#{orientation}", ".jpg"])
-    oriented_file.binmode
-    oriented_file.write(source.byteslice(0, 2) + marker + source.byteslice(2..))
-    oriented_file.rewind
-
-    yield oriented_file
-  ensure
-    oriented_file&.close!
-  end
-
-  def stored_color_grid(upload:, rows:, columns:, palette:)
-    stored_path = Discourse.store.path_for(upload)
-
-    Dir.mktmpdir do |directory|
-      png_path = File.join(directory, "stored.png")
-      ImageMagick.magick(stored_path, png_path, read: [stored_path], write: [directory])
-      stored_image = ChunkyPNG::Image.from_file(png_path)
-
-      Array.new(rows) do |row_index|
-        Array.new(columns) do |column_index|
-          x = (column_index * stored_image.width + stored_image.width / 2) / columns
-          y = (row_index * stored_image.height + stored_image.height / 2) / rows
-
-          nearest_palette_color(pixel: stored_image[x, y], palette: palette)
-        end
-      end
-    end
-  end
-
-  def nearest_palette_color(pixel:, palette:)
-    actual_rgb = [ChunkyPNG::Color.r(pixel), ChunkyPNG::Color.g(pixel), ChunkyPNG::Color.b(pixel)]
-
-    palette
-      .min_by do |_name, expected_rgb|
-        expected_rgb
-          .zip(actual_rgb)
-          .sum { |expected_channel, actual_channel| (expected_channel - actual_channel)**2 }
-      end
-      .first
-  end
 
   describe "#create_for" do
     context "when the upload is an SVG" do
@@ -259,9 +211,11 @@ RSpec.describe UploadCreator do
     end
 
     context "when optimization is forced for an animated GIF" do
-      # Regenerate from the repository root with:
-      # ruby -rbase64 -e 'File.binwrite("spec/fixtures/images/tiny_animated.gif", Base64.strict_decode64("R0lGODlhAgABAPAAAP8AAAAAACH/C05FVFNDQVBFMi4wAwEAAAAh+QQAAAAAACwAAAAAAgABAAACAgQKACH5BAAKAAAALAAAAAACAAEAgAAA/wAAAAICBAoAOw=="))'
-      let(:file) { file_from_fixtures("tiny_animated.gif") }
+      let(:file) do
+        # Regenerate from the repository root with:
+        # ruby -rbase64 -e 'File.binwrite("spec/fixtures/images/tiny_animated.gif", Base64.strict_decode64("R0lGODlhAgABAPAAAP8AAAAAACH/C05FVFNDQVBFMi4wAwEAAAAh+QQAAAAAACwAAAAAAgABAAACAgQKACH5BAAKAAAALAAAAAACAAEAgAAA/wAAAAICBAoAOw=="))'
+        file_from_fixtures("tiny_animated.gif")
+      end
 
       it "preserves the GIF and records its animation metadata" do
         upload =
@@ -279,6 +233,8 @@ RSpec.describe UploadCreator do
     end
 
     context "when a JPEG has EXIF orientation metadata" do
+      include ImageOrientationHelpers
+
       # Regenerate from the repository root:
       # magick \( xc:red xc:lime xc:blue +append \) \( xc:cyan xc:magenta xc:yellow +append \) -append -scale 60x40! -sampling-factor 4:4:4 -quality 95 spec/fixtures/images/exif_orientation.jpg
       let(:source_path) { Rails.root.join("spec/fixtures/images/exif_orientation.jpg") }
